python-code/pyproject.toml
[tool.poetry]
name = "updater"
version = "1.0.7"
description = ""
authors = ["Ru Chern Chong <ruchern.chong@sgcarstrends.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
arpeggio = "2.0.2"
attrs = "23.2.0"
black = "24.4.2"
certifi = "2024.7.4"
charset-normalizer = "3.3.2"
click = "8.1.7"
dill = "0.3.8"
dnspython = "2.6.1"
grpcio = "1.60.1"
idna = "3.7"
iniconfig = "2.0.0"
mypy-extensions = "1.0.0"
packaging = "24.0"
parver = "0.5"
pathspec = "0.12.1"
platformdirs = "4.2.2"
pluggy = "1.5.0"
protobuf = "4.25.3"
pulumi = "3.116.1"
pulumi-aws = "6.36.0"
pymongo = "^4.9.1"
pytest = "8.2.0"
python-dotenv = "1.0.1"
pyyaml = "6.0.1"
requests = "2.32.2"
semver = "2.13.0"
six = "1.16.0"
urllib3 = "2.2.2"
pre-commit = "^3.8.0"
upstash-redis = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
